feature: added attendee entity
+ added Event-Attendee relation in entity

+++ src/events/events.controller.ts
@@ -5,6 +5,7 @@ import {
   Get,
   HttpCode,
   Logger,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -12,7 +13,7 @@ import {
 import { CreateEventDto } from './create-event.dto';
 import { UpdateEventDto } from './update-event.dto';
 
-import { Like, MoreThan, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Event } from './event.entity';
 
@@ -27,32 +28,38 @@ export class EventsController {
 
   @Get()
   async findAll() {
-    this.logger.log(`Hit the findAll route`);
+    // this.logger.log(`Hit the findAll route`);
     const events = await this.repository.find();
-    this.logger.debug(`Found ${events.length} events`);
+    // this.logger.debug(`Found ${events.length} events`);
     return events;
   }
 
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      where: [
-        {
-          id: MoreThan(3),
-          when: MoreThan(new Date('2021-02-12T13:00:00')),
-        },
-        { description: Like('%meet%') },
-      ],
-      take: 2,
-      order: {
-        id: 'DESC',
-      },
-    });
-  }
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repository.find({
+  //     where: [
+  //       {
+  //         id: MoreThan(3),
+  //         when: MoreThan(new Date('2021-02-12T13:00:00')),
+  //       },
+  //       { description: Like('%meet%') },
+  //     ],
+  //     take: 2,
+  //     order: {
+  //       id: 'DESC',
+  //     },
+  //   });
+  // }
 
   @Get(':id')
   async findOne(@Param('id') id) {
-    return await this.repository.findOne(id);
+    const event = await this.repository.findOneBy({ id });
+
+    if (!event) {
+      throw new NotFoundException();
+    }
+
+    return event;
   }
 
   @Post()
@@ -65,7 +72,11 @@ export class EventsController {
 
   @Patch(':id')
   async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOneBy({ id });
+
+    if (!event) {
+      throw new NotFoundException();
+    }
 
     return await this.repository.save({
       ...event,
@@ -77,7 +88,11 @@ export class EventsController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id') id) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOneBy({ id });
+
+    if (!event) {
+      throw new NotFoundException();
+    }
 
     await this.repository.remove(event);
   }
